+++ Samples/MSRP/MsrpServer/Program.cs
@@ -71,7 +71,7 @@ internal class Program
 
     private static void OnTextMessageReceived(string message, string from)
     {
-        Console.WriteLine(message.Replace("\r\n", ""));
+        Console.WriteLine($"From {from}: {message.Replace("\r\n", "")}");
     }
 
     private static void OnInviteReceived(string from)