+++ VersionHistory.md
@@ -5,3 +5,12 @@
 |--------|--------|-------|
 | NA       |  New      | Initial version |
 
+## v0.0.2 - 16 Sep 2024
+| Issue No. | Change Type | Description |
+|--------|--------|-------|
+| NA       |  Change  | Added a "user" parameter to SipUtils.CreateMsrpMediaDescription(). Changed the setupType to be required. |
+| NA     | Change | Removed the OfferedSdp and AnsweredSdp parameters from MsrpConnection.CreateFromSdp() because they are not used. |
+| NA      | Change | SipLib.Msrp.MsrpConnection -- Added a new private method calld SendEmptySendRequest(). |
+| NA      | Change | SipLib.Msrp.MessageReceivedDelegate -- added the "from" parameter. |
+
+